add check for required stopping criterion & evaluation count is now  long

+++ src/experiments/permus/PERMU_params.h
@@ -13,7 +13,7 @@ namespace PERMU{
 
         // PARAMETERS TO BE SPECIFIED IN CONFIG FILE//
         double MAX_SOLVER_TIME;
-        int MAX_SOLVER_FE;
+        long MAX_SOLVER_FE;
         int POPSIZE;
         int N_EVALS;
         int TABU_LENGTH;
